requirements.txt
django==4.2
django-phonenumber-field[phonenumbers]
Pillow==9.5.0